test: split the two long counting test files by concern
Pure move along existing section banners: test_counted_float.py splits into value/behavior vs flop-counting, and test_math_patching.py into patch machinery vs per-function counting, with the shared PATCHED_FUNCTION_NAMES / STDLIB_MATH_FUNCTIONS constants lifted to conftest so both halves reach them. No test added, removed or changed; the suite collects and passes identically.
